add flags -h/-c
Split main

cobra library for flags
